# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Google Scraper API Documentation",
        "description": "API documentation for Google Scraper application",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/keywords": {
            "get": {
                "tags": [
                    "Keywords"
                ],
                "summary": "Get all keywords",
                "description": "Returns a list of all keywords for the authenticated user",
                "operationId": "getKeywords",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "keyword": {
                                                "type": "string",
                                                "example": "cloud computing"
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": [
                                                    "pending",
                                                    "processing",
                                                    "completed",
                                                    "failed"
                                                ],
                                                "example": "completed"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/keywords/upload": {
            "post": {
                "tags": [
                    "Keywords"
                ],
                "summary": "Upload keywords",
                "description": "Upload a CSV file containing keywords to scrape",
                "operationId": "uploadKeywords",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "CSV file containing keywords (max 1MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Keywords uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "5 keywords uploaded successfully"
                                        },
                                        "keywords": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "keyword": {
                                                        "type": "string",
                                                        "example": "cloud computing"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "file": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Maximum 100 keywords allowed per upload. Found: 101 keywords."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/keywords/{keyword}": {
            "get": {
                "tags": [
                    "Keywords"
                ],
                "summary": "Get keyword details",
                "description": "Returns details of a specific keyword",
                "operationId": "getKeyword",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "path",
                        "description": "Keyword ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "keyword": {
                                            "type": "string",
                                            "example": "cloud computing"
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "pending",
                                                "processing",
                                                "completed",
                                                "failed"
                                            ],
                                            "example": "completed"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Keyword not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/keywords/{keyword}/retry": {
            "post": {
                "tags": [
                    "Keywords"
                ],
                "summary": "Retry failed keyword",
                "description": "Retries scraping for a failed keyword",
                "operationId": "retryKeyword",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "path",
                        "description": "Keyword ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Keyword queued for retry",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Keyword queued for retry"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Keyword not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/search-results/{keywordId}": {
            "get": {
                "tags": [
                    "Search Results"
                ],
                "summary": "Get search results",
                "description": "Returns the latest search results for a specific keyword",
                "operationId": "getSearchResults",
                "parameters": [
                    {
                        "name": "keywordId",
                        "in": "path",
                        "description": "Keyword ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "keyword_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total_ads": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "total_links": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "organic_results": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Cloud Computing Services - AWS"
                                                    },
                                                    "url": {
                                                        "type": "string",
                                                        "example": "https://aws.amazon.com"
                                                    },
                                                    "snippet": {
                                                        "type": "string",
                                                        "example": "Amazon Web Services offers reliable, scalable, and inexpensive cloud computing services."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "completed",
                                                "failed"
                                            ],
                                            "example": "completed"
                                        },
                                        "error_message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "scraped_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Keyword not found or no search results available"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "tags": [
        {
            "name": "Keywords",
            "description": "API Endpoints for managing keywords"
        },
        {
            "name": "Search Results",
            "description": "API Endpoints for managing search results"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ]
}